refactor: FollowDto 레코드 위치 변경 취소
