+++ src/main/java/sws/songpin/domain/follow/dto/response/FollowDto.java
@@ -1,4 +1,4 @@
-package sws.songpin.domain.follow.dto;
+package sws.songpin.domain.follow.dto.response;
 
 import sws.songpin.domain.member.entity.ProfileImg;
