+++ src/main/java/sws/songpin/domain/follow/dto/response/FollowerListResponseDto.java
@@ -1,7 +1,5 @@
 package sws.songpin.domain.follow.dto.response;
 
-import sws.songpin.domain.follow.dto.FollowDto;
-
 import java.util.List;
 
 public record FollowerListResponseDto(
